# src/main/prober/web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only browser views over the prober's analysis engine, adapted from `ProbeSession` \u2014 the same facade `python -m main.prober.query` uses. No analysis lives here. Reading is anonymous; the work is not: the background probes and every route that loads a checkpoint or runs the model require the shared password.",
    "title": "gen3ai prober",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/game/attention": {
      "get": {
        "operationId": "api_game_attention_api_game_attention_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          },
          {
            "in": "query",
            "name": "inv",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Inv",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          },
          {
            "in": "query",
            "name": "head",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Head"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Game Attention Api Game Attention Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "locked: needs the shared password (/login)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One decision's full attention map (query \u00d7 key), one layer/head or their average",
        "tags": [
          "read-only"
        ]
      }
    },
    "/api/game/board": {
      "get": {
        "operationId": "api_game_board_api_game_board_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          },
          {
            "in": "query",
            "name": "inv",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Inv",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Game Board Api Game Board Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The board one decision was made on, every fact tagged public / ours / hidden (model-free)",
        "tags": [
          "read-only"
        ]
      }
    },
    "/api/game/readout": {
      "get": {
        "operationId": "api_game_readout_api_game_readout_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Game Readout Api Game Readout Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "locked: needs the shared password (/login)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One battle's model panels: intent, beliefs, attention summary, pointer scores, operator facts (LOADS the checkpoint; current architecture only)",
        "tags": [
          "read-only"
        ]
      }
    },
    "/api/game/story": {
      "get": {
        "operationId": "api_game_story_api_game_story_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Game Story Api Game Story Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One battle's turn story: per turn the protocol events, the ordered beats, the board; the decisions (model-free)",
        "tags": [
          "read-only"
        ]
      }
    },
...
    "/battle": {
      "get": {
        "operationId": "page_battle_redirect_battle_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          },
          {
            "description": "game turn (mapped to turn=)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "game turn (mapped to turn=)",
              "title": "Start"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The classic replay, MERGED into /game (2026-10-09): redirects there, keeping the battle and mapping start=N to turn=N",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/game": {
      "get": {
        "operationId": "page_game_game_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          },
          {
            "description": "decision index; default = the first",
            "in": "query",
            "name": "inv",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "decision index; default = the first",
              "title": "Inv"
            }
          },
          {
            "description": "game turn; selects its first decision",
            "in": "query",
            "name": "turn",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "game turn; selects its first decision",
              "title": "Turn"
            }
          },
          {
            "description": "perspective: model (default) | truth | public",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "perspective: model (default) | truth | public",
              "title": "View"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The battle viewer: the turn story, the board under an information perspective, our options vs their action, and (unlocked) the model's beliefs, intent and attention",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/partials/game/attention": {
      "get": {
        "operationId": "partial_game_attention_partials_game_attention_get",
        "parameters": [
          {
            "in": "query",
            "name": "run",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run"
            }
          },
          {
            "in": "query",
            "name": "battle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battle"
            }
          },
          {
            "in": "query",
            "name": "inv",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "0",
              "title": "Inv"
            }
          },
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          },
          {
            "in": "query",
            "name": "head",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Head"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "locked: needs the shared password (/login)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "/game's attention heat map for one decision (HTMX target)",
        "tags": [
          "partials"
        ]
      }
    },
...
  }
}